Fixed GetPercentCorrect
Cast numberCorrectAnswers and numberOfQuestions to doubles.

+++ PrincessBrideTrivia/PrincessBrideTrivia/Program.cs
@@ -21,7 +21,7 @@ public class Program
 
     public static string GetPercentCorrect(int numberCorrectAnswers, int numberOfQuestions)
     {
-        return (numberCorrectAnswers / numberOfQuestions * 100) + "%";
+        return ((double)numberCorrectAnswers / (double)numberOfQuestions * 100) + "%";
     }
 
     public static bool AskQuestion(Question question)
